requirements.txt
pyserial
paho-mqtt
cryptography